Split stage artefact BDD tests from pipeline tests
Move shared stage-release-artefacts test path setup and the BDD context
fixture into `conftest.py`, then move the cargo-binstall behaviour steps
out of `test_pipeline.py` into a dedicated BDD module.

Remove repeated script path bootstrapping from the focused test modules
so each file mirrors its source module more closely.
