+++ .github/actions/stage-release-artefacts/tests/test_config.py
@@ -2,14 +2,9 @@
 
 from __future__ import annotations
 
-from pathlib import Path
+import typing as typ
 
 import pytest
-from syspath_hack import prepend_to_syspath
-
-SCRIPTS_DIR = Path(__file__).resolve().parent.parent / "scripts"
-prepend_to_syspath(SCRIPTS_DIR)
-
 from stage_common import StageError
 from stage_common.config import (
     ArtefactConfig,
@@ -17,6 +12,9 @@ from stage_common.config import (
     StagingConfig,
     load_config,
 )
+
+if typ.TYPE_CHECKING:
+    from pathlib import Path
 
 
 class TestArtefactConfig:
